requirements.txt
psycopg2-binary
pydantic-settings
python-dotenv
fastapi
uvicorn[standard]
email-validator
sqlalchemy
requests

